move find free port to rest server extension
