+++ open-api/src/testFixtures/java/org/apache/iceberg/rest/RCKUtils.java
@@ -18,6 +18,9 @@
  */
 package org.apache.iceberg.rest;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
+import java.net.ServerSocket;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
@@ -113,5 +116,13 @@ class RCKUtils {
               catalog.listViews(namespace).forEach(catalog::dropView);
               catalog.dropNamespace(namespace);
             });
+  }
+
+  static int findFreePort() {
+    try (ServerSocket socket = new ServerSocket(0)) {
+      return socket.getLocalPort();
+    } catch (IOException e) {
+      throw new UncheckedIOException(e);
+    }
   }
 }
